add configuration file

+++ src/tasks/tasks.service.ts
@@ -14,12 +14,12 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  getTasks(taskFilterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(taskFilterDto);
+  getTasks(taskFilterDto: GetTaskFilterDto, user:User): Promise<Task[]> {
+    return this.taskRepository.getTasks(taskFilterDto, user);
   }
 
-  async getTaskByID(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskByID(id: number, user:User): Promise<Task> {
+    const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
     if (!found) {
       throw new NotFoundException(`Task with ID "${id} not found"`);
     }
@@ -30,15 +30,15 @@ export class TasksService {
     return this.taskRepository.createTask(createTask, user);
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskByID(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskByID(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id} not found"`);
